NOWSEOUL-37: Integrate geoJSON for advanced mapping and implement carousel view
This commit introduces significant advancements in handling and visualizing geographic data within the app, leveraging geoJSON for detailed mapping of Seoul's hotspot areas. Key components include the definition of new data models for geoJSON coordinates, the addition of geoJSON files for Seoul's hotspot areas, enhancements to MapView for displaying these areas, and the implementation of a carousel view for intuitive navigation and presentation.

- Defined new MDF models (Feature, HotspotArea, etc.) to accurately represent geoJSON coordinate data
- Added geoJSON files ('seoul_hotspots_113.geojson' and 'seoul_hotspots_113.json') to the project, enabling precise mapping of Seoul's hotspot areas.
- Enhanced the MapView and MapViewController to integrate geoJSON data, allowing for dynamic rendering of polygons representing Seoul's regions on the map.
- Implemented a carousel view in the MapScene, offering users an interactive and engaging way to explore information about hotspot areas.